requirements.txt
streamlit
google-generativeai
python-dotenv
gTTS
streamlit-mic-recorder
openai-whisper
ffmpeg-python